add direct connect with optional initial payload

+++ include/bbclient/bb_connection.h
@@ -64,7 +64,7 @@ b32 bbcon_is_connected(const bb_connection_t *con);
 void bbcon_disconnect(bb_connection_t *con);
 
 void bbcon_send(bb_connection_t *con, bb_decoded_packet_t *decoded);
-void bbcon_send_raw(bb_connection_t *con, const void *pData, u16 nBytes);
+void bbcon_send_raw(bb_connection_t *con, const void *pData, u32 nBytes);
 void bbcon_flush(bb_connection_t *con);
 
 b32 bbcon_try_send(bb_connection_t *con, bb_decoded_packet_t *decoded); // returns false if send would block
